+++ mmodel/__init__.py
@@ -1,4 +1,3 @@
 from mmodel.graph import *
 from mmodel.handler import *
 from mmodel.model import *
-from mmodel.modifier import *